+++ src/components/ChatNavbar/ChatNavbars.jsx
@@ -116,7 +116,7 @@ const ChatNavbars = ({ onTweetSubmit }) => {
           </Link>
           <button className={style.NavbarButton} onClick={handlebuttonClick}>推文</button>
         </div>
-        {openModal && <TweetModal open={openModal} onClose={() => setOpenModal(false)} onTweetSubmit={onTweetSubmit} />}
+        {openModal && <TweetModal open={openModal} onClose={(res) => setOpenModal(res)} onTweetSubmit={onTweetSubmit} />}
         <div className={style.logout} onClick={handleLogout}><img className={style.NavbarPng} src={LogoutIcon} alt="logout" />登出</div>
       </div>
